requirements.txt
django
Pillow
mysql-python
uwsgi
django-ckeditor